Refactored game state for legibility
