+++ server/pyproject.toml
@@ -3,7 +3,7 @@ name = "server"
 version = "0.1.0"
 description = "Add your description here"
 readme = "README.md"
-requires-python = ">=3.12"
+requires-python = "==3.12"
 dependencies = [
     "black>=24.8.0",
     "fastapi[standard]>=0.112.2",
